# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MakeYourChoice IU API
  version: 1.0.0
  description: API для системы выбора элективов Иннополиса
servers:
  - url: http://localhost:8000/api
    description: Local development
...
paths:
...
  /electives/{id}:
    get:
      tags: [Electives]
      summary: Получить электив по id
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Elective"
          description: Электив
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [Electives]
      summary: Изменить поля электива (включая удаление через status -1)
      description: >
        Удаление реализовано через PATCH — передаём `{"status": -1}`.
        Возвращает `{"status": "archived"}`.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ElectiveUpdate"
            examples:
              update_fields:
                summary: Изменить поля
                value:
                  elective_language: russian
                  prerequisite: Example prerequisite
                  status: 1
              delete:
                summary: Удалить (мягкое удаление)
                value:
                  status: -1
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    id:
      in: path
      name: id
      required: true
      schema:
        type: integer
  responses:
    Success:
      description: Операция выполнена успешно
      content:
        application/json:
          example:
            status: success
    Error:
      description: Ошибка валидации или бизнес-логики
      content:
        application/json:
          example:
            status: error
    NotFound:
      description: Объект не найден
      content:
        application/json:
          example:
            status: error
            message: Not found
...
  schemas:
    Elective:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        instructor:
          type: string
        description:
          type: string
        elective_type:
          type: string
          nullable: true
          example: TECH
        program_language:
          type: string
          example: ENG
        elective_language:
          type: string
          example: english
        status:
          type: integer
          description: "0 — архив, 1 — активный, -1 — удалён"
        degree_year:
          type: array
          items:
            type: string
          example: ["BS-1", "BS-2"]
        prerequisite:
          type: string
...
    ElectiveUpdate:
      type: object
      properties:
        name:
          type: string
        instructor:
          type: string
        description:
          type: string
        elective_language:
          type: string
        program_language:
          type: string
        elective_type:
          type: string
        status:
          type: integer
          description: "0 — архивировать, 1 — активировать, -1 — удалить"
        degree_year:
          type: array
          items:
            type: string
        prerequisite:
          type: string
...